Avoid cloning entire cache fields by adding ReadOnlyMapRef
